requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0

# LLM Providers
anthropic>=0.18.0
openai>=1.0.0
groq>=0.4.0
together>=0.2.0

# Local LLM
ollama>=0.1.0

# Embeddings
sentence-transformers>=2.2.0
fastembed>=0.1.0

# Database
psycopg2-binary>=2.9.0
redis>=5.0.0
mysql-connector-python>=8.0.0
motor>=3.3.0

# Memory & Search
faiss-cpu>=1.7.4
chromadb>=0.4.0
rank-bm25>=0.2.2
cohere>=5.0.0
numpy>=1.24.0


# Monitoring
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Utils
pydantic>=2.0.0
tenacity>=8.2.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
